# src/main/java/com/abnamro/recipes/web/dto/ProblemDto.java
package com.abnamro.recipes.web.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.time.LocalDateTime;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class ProblemDto {

    private LocalDateTime dateTime;
    private String error;
}